requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
flask
schedule
python-dotenv
gunicorn
browser-use
langchain-google-genai
playwright
